refactor: add router plugin
