+++ packages/nuxt/src/pages/runtime/plugins/router-typed.ts
@@ -1,0 +1,9 @@
+import { createRouter } from 'vue-router/auto'
+import { createRouterPlugin } from './router-base'
+
+export default createRouterPlugin((history, routerOptions) => {
+  return createRouter({
+    ...routerOptions,
+    history
+  })
+})
